requirements.txt
torch>=1.12.1
transformers>=4.21.0
datasets>=2.5.1
bertopic>=0.14.0
farm-haystack>=1.14.0
batterydataextractor>=0.0.5
summa>=1.2.0
numba==0.54.1
rouge>=1.0.1
